chore: refactor program fields in their own file
